Improve gem tabs UI and fix a small UI bug with updating active item in item picker
