+++ ui/core/_shared.scss
@@ -66,10 +66,6 @@
   100% { transform: rotate(360deg); }
 }
 
-.socket-color-blue {
-  border: 1px solid blue;
-}
-
 .item-quality-junk {
   color: #fff !important;
 }
@@ -94,16 +90,20 @@
   color: #ff8000 !important;
 }
 
+.socket-color-blue {
+  border: 1px solid blue !important;
+}
+
 .socket-color-meta {
-  border: 1px solid white;
+  border: 1px solid white !important;
 }
 
 .socket-color-red {
-  border: 1px solid red;
+  border: 1px solid red !important;
 }
 
 .socket-color-yellow {
-  border: 1px solid yellow;
+  border: 1px solid yellow !important;
 }
 
 @mixin wowhead-background-icon {
